+++ nf_core/__main__.py
@@ -371,10 +371,9 @@ def list(ctx, pipeline_dir, json):
     If no pipeline directory is given, lists all currently available
     software wrappers in the nf-core/modules repository.
     """
-    mods = nf_core.modules.PipelineModules()
-    mods.modules_repo = ctx.obj["modules_repo_obj"]
-    mods.pipeline_dir = pipeline_dir
-    print(mods.list_modules(json))
+    module_list = nf_core.modules.ModuleList(pipeline_dir)
+    module_list.modules_repo = ctx.obj["modules_repo_obj"]
+    print(module_list.list_modules(json))
 
 
 @modules.command(help_priority=2)
@@ -394,13 +393,9 @@ def install(ctx, pipeline_dir, tool, latest, force, sha):
     along with associated metadata.
     """
     try:
-        mods = nf_core.modules.PipelineModules()
-        mods.modules_repo = ctx.obj["modules_repo_obj"]
-        mods.pipeline_dir = pipeline_dir
-        mods.force = force
-        mods.latest = latest
-        mods.sha = sha
-        mods.install(tool)
+        module_install = nf_core.modules.ModuleInstall(pipeline_dir, force=force, latest=latest, sha=sha)
+        module_install.modules_repo = ctx.obj["modules_repo_obj"]
+        module_install.install(tool)
     except UserWarning as e:
         log.critical(e)
         sys.exit(1)
@@ -415,10 +410,9 @@ def remove(ctx, pipeline_dir, tool):
     Remove a software wrapper from a pipeline.
     """
     try:
-        mods = nf_core.modules.PipelineModules()
-        mods.modules_repo = ctx.obj["modules_repo_obj"]
-        mods.pipeline_dir = pipeline_dir
-        mods.remove(tool)
+        module_remove = nf_core.modules.ModuleRemove(pipeline_dir)
+        module_remove.modules_repo = ctx.obj["modules_repo_obj"]
+        module_remove.remove(tool)
     except UserWarning as e:
         log.critical(e)
         sys.exit(1)